requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
# Azure Services
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-ai-formrecognizer>=3.3.0
azure-storage-file-datalake>=12.14.0

# Azure OpenAI
openai>=1.12.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
